WIP on different scene transitions
