+++ Assets/Scenes/Client/ClientLobby.unity
@@ -203,7 +203,7 @@ MonoBehaviour:
   m_OnClick:
     m_PersistentCalls:
       m_Calls:
-      - m_Target: {fileID: 1668815429}
+      - m_Target: {fileID: 0}
         m_MethodName: 
         m_Mode: 1
         m_Arguments:
@@ -251,30 +251,6 @@ CanvasRenderer:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 299737380}
   m_CullTransparentMesh: 0
---- !u!114 &999353107 stripped
-MonoBehaviour:
-  m_CorrespondingSourceObject: {fileID: 1584834601242082865, guid: c732e1b58c017874bbf5da45531902cf,
-    type: 3}
-  m_PrefabInstance: {fileID: 1584834602100937690}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 56f55256fc4a6704eafd0dd78732d19e, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
---- !u!114 &1004355448 stripped
-MonoBehaviour:
-  m_CorrespondingSourceObject: {fileID: 5214724981417358588, guid: c732e1b58c017874bbf5da45531902cf,
-    type: 3}
-  m_PrefabInstance: {fileID: 1584834602100937690}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 37271d27156413948aba6f3d52b03edf, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
 --- !u!1 &1216002073
 GameObject:
   m_ObjectHideFlags: 0
@@ -672,18 +648,6 @@ Transform:
   m_Father: {fileID: 0}
   m_RootOrder: 2
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!114 &1668815429 stripped
-MonoBehaviour:
-  m_CorrespondingSourceObject: {fileID: 1584834600499235231, guid: c732e1b58c017874bbf5da45531902cf,
-    type: 3}
-  m_PrefabInstance: {fileID: 1584834602100937690}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: ad6e58ddc91bc2e41a1691bec489cc39, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
 --- !u!1 &2021493589
 GameObject:
   m_ObjectHideFlags: 0
@@ -766,7 +730,7 @@ MonoBehaviour:
   m_OnClick:
     m_PersistentCalls:
       m_Calls:
-      - m_Target: {fileID: 1668815429}
+      - m_Target: {fileID: 0}
         m_MethodName: 
         m_Mode: 1
         m_Arguments:
@@ -814,100 +778,6 @@ CanvasRenderer:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 2021493589}
   m_CullTransparentMesh: 0
---- !u!1001 &1584834602100937690
-PrefabInstance:
-  m_ObjectHideFlags: 0
-  serializedVersion: 2
-  m_Modification:
-    m_TransformParent: {fileID: 0}
-    m_Modifications:
-    - target: {fileID: 1584834600499235231, guid: c732e1b58c017874bbf5da45531902cf,
-        type: 3}
-      propertyPath: m_UDPClient
-      value: 
-      objectReference: {fileID: 999353107}
-    - target: {fileID: 1584834600499235231, guid: c732e1b58c017874bbf5da45531902cf,
-        type: 3}
-      propertyPath: m_TCPClient
-      value: 
-      objectReference: {fileID: 1004355448}
-    - target: {fileID: 1584834600499235231, guid: c732e1b58c017874bbf5da45531902cf,
-        type: 3}
-      propertyPath: m_worldGrid
-      value: 
-      objectReference: {fileID: 0}
-    - target: {fileID: 1584834600992003656, guid: c732e1b58c017874bbf5da45531902cf,
-        type: 3}
-      propertyPath: m_LocalPosition.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 1584834600992003656, guid: c732e1b58c017874bbf5da45531902cf,
-        type: 3}
-      propertyPath: m_LocalPosition.y
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 1584834600992003656, guid: c732e1b58c017874bbf5da45531902cf,
-        type: 3}
-      propertyPath: m_LocalPosition.z
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 1584834600992003656, guid: c732e1b58c017874bbf5da45531902cf,
-        type: 3}
-      propertyPath: m_LocalRotation.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 1584834600992003656, guid: c732e1b58c017874bbf5da45531902cf,
-        type: 3}
-      propertyPath: m_LocalRotation.y
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 1584834600992003656, guid: c732e1b58c017874bbf5da45531902cf,
-        type: 3}
-      propertyPath: m_LocalRotation.z
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 1584834600992003656, guid: c732e1b58c017874bbf5da45531902cf,
-        type: 3}
-      propertyPath: m_LocalRotation.w
-      value: 1
-      objectReference: {fileID: 0}
-    - target: {fileID: 1584834600992003656, guid: c732e1b58c017874bbf5da45531902cf,
-        type: 3}
-      propertyPath: m_RootOrder
-      value: 3
-      objectReference: {fileID: 0}
-    - target: {fileID: 1584834600992003656, guid: c732e1b58c017874bbf5da45531902cf,
-        type: 3}
-      propertyPath: m_LocalEulerAnglesHint.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 1584834600992003656, guid: c732e1b58c017874bbf5da45531902cf,
-        type: 3}
-      propertyPath: m_LocalEulerAnglesHint.y
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 1584834600992003656, guid: c732e1b58c017874bbf5da45531902cf,
-        type: 3}
-      propertyPath: m_LocalEulerAnglesHint.z
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 1584834600992003659, guid: c732e1b58c017874bbf5da45531902cf,
-        type: 3}
-      propertyPath: m_Name
-      value: ClientNetworking
-      objectReference: {fileID: 0}
-    - target: {fileID: 1584834601242082864, guid: c732e1b58c017874bbf5da45531902cf,
-        type: 3}
-      propertyPath: m_Name
-      value: UDPClient
-      objectReference: {fileID: 0}
-    - target: {fileID: 8752211695502532364, guid: c732e1b58c017874bbf5da45531902cf,
-        type: 3}
-      propertyPath: m_Name
-      value: TCPClient
-      objectReference: {fileID: 0}
-    m_RemovedComponents: []
-  m_SourcePrefab: {fileID: 100100000, guid: c732e1b58c017874bbf5da45531902cf, type: 3}
 --- !u!1001 &4639765598835359790
 PrefabInstance:
   m_ObjectHideFlags: 0
