+++ Assets/Scripts/Client/ClientSyncStates.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0529b9dc3c412de4a9dea0d4d2bb5ed8
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
