+++ Assets/Scripts/Client/ClientSyncStates/ClientSyncInit.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f40a56306a6e45b41bbae31a0f3a7f51
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
