+++ Assets/Scripts/Client/ClientSyncStates/ClientSyncLoadWorld.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: af48666a0ead02640af7a30caf87caca
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
