+++ Assets/Scripts/Client/ClientSyncStates/ClientSyncPlay.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: dd061c09ef56aae4aad8621ac5bd44f0
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
